Wait for page change before matching content
