+++ spec/spec_helper.rb
@@ -6,7 +6,7 @@ require "webmock/rspec"
 require "paperclip/matchers"
 require "active_storage_validations/matchers"
 require "./spec/deprecation_toolkit_env"
-require "./spec/support/wait_for_ajax"
+require "./spec/support/wait_for_helpers"
 require "./spec/support/select_from_chosen"
 
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
@@ -103,7 +103,7 @@ RSpec.configure do |config|
 
   config.include ActiveSupport::Testing::TimeHelpers
   config.include TimeTravelHelpers
-  config.include WaitForAjax
+  config.include WaitForHelpers
   config.include SelectFromChosen
 
   config.around(:each, :time_travel) do |example|
